# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/adminproductdemo": {
      "post": {
        "operationId": "AdminproductdemoController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adminproductdemo"
                }
              }
            }
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      },
      "get": {
        "operationId": "AdminproductdemoController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      }
    },
...
    "/adminproductdemo/{id}": {
      "get": {
        "operationId": "AdminproductdemoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adminproductdemo"
                }
              }
            }
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      },
      "patch": {
        "operationId": "AdminproductdemoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adminproductdemo"
                }
              }
            }
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      },
      "delete": {
        "operationId": "AdminproductdemoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      }
    },
    "/adminproductdemo/{id}/updates": {
      "patch": {
        "operationId": "AdminproductdemoController_updates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminproductdemoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adminproductdemo"
                }
              }
            }
          }
        },
        "tags": [
          "Adminproductdemo"
        ]
      }
    },
    "/announcement": {
      "post": {
        "operationId": "AnnouncementController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Announcement"
                }
              }
            }
          }
        },
        "tags": [
          "Announcement"
        ]
      },
      "get": {
        "operationId": "AnnouncementController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
...
    "/announcement/{id}": {
      "get": {
        "operationId": "AnnouncementController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Announcement"
                }
              }
            }
          }
        },
        "tags": [
          "Announcement"
        ]
      },
      "patch": {
        "operationId": "AnnouncementController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Announcement"
                }
              }
            }
          }
        },
        "tags": [
          "Announcement"
        ]
      },
      "delete": {
        "operationId": "AnnouncementController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
    "/announcement/{announcementId}/read": {
      "post": {
        "operationId": "AnnouncementController_markRead",
        "parameters": [
          {
            "name": "announcementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
    "/announcement/{announcementId}/unread": {
      "delete": {
        "operationId": "AnnouncementController_markAsUnread",
        "parameters": [
          {
            "name": "announcementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
    "/announcement/{announcementId}/readers": {
      "get": {
        "operationId": "AnnouncementController_getReaders",
        "parameters": [
          {
            "name": "announcementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementRead"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
    "/announcement/{announcementId}/read-stats": {
      "get": {
        "operationId": "AnnouncementController_getReadStats",
        "parameters": [
          {
            "name": "announcementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Announcement"
        ]
      }
    },
...
  },
  "info": {
    "title": "CN Squad API",
    "description": "The CN Squad API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "city",
          "state",
          "country",
          "createdAt",
          "updatedAt"
        ]
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nameAbrv": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nameAbrv",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "DepartmentalRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "department",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffEmployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "employeeCode": {
            "type": "string"
          },
          "jobTitle": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "Full-Time",
              "Part-Time",
              "Contract",
              "Intern"
            ]
          },
          "workMode": {
            "type": "string",
            "enum": [
              "Onsite",
              "Remote",
              "Hybrid"
            ]
          },
          "hireDate": {
            "type": "string"
          },
          "reportingManager": {
            "type": "string"
          },
          "directReport": {
            "type": "string"
          },
          "contractEndDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended",
              "Terminated"
            ]
          },
          "workLocation": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "departmentalRole": {
            "$ref": "#/components/schemas/DepartmentalRole"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "employeeCode",
          "jobTitle",
          "employmentType",
          "workMode",
          "hireDate",
          "reportingManager",
          "directReport",
          "status",
          "department",
          "departmentalRole",
          "createdAt",
          "updatedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "staff": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Staff"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "permissions",
          "staff",
          "createdAt",
          "updatedAt"
        ]
      },
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hobbies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "Male",
              "Female",
              "Other"
            ],
            "type": "string"
          },
          "maritalStatus": {
            "enum": [
              "Single",
              "Married",
              "Divorced",
              "Widowed"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/StaffAddress"
          },
          "employment": {
            "$ref": "#/components/schemas/StaffEmployment"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "resetToken": {
            "type": "string"
          },
          "resetTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "registrationToken": {
            "type": "string"
          },
          "registrationTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "isRegistered": {
            "type": "boolean"
          },
          "lastIpAddress": {
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isCeo": {
            "type": "boolean"
          },
          "isOnline": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "firstName",
          "lastName",
          "description",
          "hobbies",
          "dateOfBirth",
          "gender",
          "email",
          "phone",
          "password",
          "address",
          "employment",
          "roles",
          "isRegistered",
          "isCeo",
          "isOnline",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isAdmin",
          "firstName",
          "lastName",
          "email",
          "staff",
          "password",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {}
      },
...
      "AnnouncementRead": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "announcement": {
            "$ref": "#/components/schemas/Announcement"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "readAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "announcement",
          "staff",
          "readAt"
        ]
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Admin"
          },
          "selectedStaffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Staff"
            }
          },
          "fileUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnnouncementRead"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "createdBy",
          "selectedStaffs",
          "fileUrls",
          "reads",
          "createdAt",
          "updatedAt"
        ]
      },
      "Adminproductdemo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nameOfProduct": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "howItWorks": {
            "type": "string"
          },
          "videos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdBy": {
            "$ref": "#/components/schemas/Admin"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nameOfProduct",
          "description",
          "howItWorks",
          "videos",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateAdminproductdemoDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}